move assignmentHandler from js_modules
