+++ src/common/components/Assignments/index.jsx
@@ -9,14 +9,14 @@ import PropTypes from 'prop-types';
 import {
   memo, useEffect, useState,
 } from 'react';
-import bc from '../../common/services/breathecode';
-import Icon from '../../common/components/Icon';
-import Text from '../../common/components/Text';
-import SimpleModal from '../../common/components/SimpleModal';
-import useStyle from '../../common/hooks/useStyle';
-import { ORIGIN_HOST } from '../../utils/variables';
-import ReviewModalComponent from '../../common/components/ReviewModal';
-import UndoApprovalModal from '../../common/components/UndoApprovalModal';
+import bc from '../../services/breathecode';
+import Icon from '../Icon';
+import Text from '../Text';
+import SimpleModal from '../SimpleModal';
+import useStyle from '../../hooks/useStyle';
+import { ORIGIN_HOST } from '../../../utils/variables';
+import ReviewModalComponent from '../ReviewModal';
+import UndoApprovalModal from '../UndoApprovalModal';
 
 export function DetailsModal({
   currentTask, projectLink, updpateAssignment, isOpen, onClose, readOnly,
